refactor(agent-bundle): move test-only host-contract smoke harnesses to tests/support
